# Optimize nonbonded precomputed