Modifies the nonbonded precomputed potential to match up with the speed optimizations made in the regular nonbonded potentials.

Improves performance of RBFE runs by about 6%

# Benchmarks
A10 card
Only focus on the RBFE tests of the benchmark as they have the nonbonded precomputed potentials

## Master
```
hif2a-rbfe: N=8840 speed: 561.40ns/day dt: 1.5fs (ran 100000 steps in 23.18s)
solvent-rbfe: N=6317 speed: 684.22ns/day dt: 1.5fs (ran 100000 steps in 19.03s)
```

## Optimized Nonbonded Precomputed
```
hif2a-rbfe: N=8840 speed: 596.80ns/day dt: 1.5fs (ran 100000 steps in 21.72s)
solvent-rbfe: N=6317 speed: 726.65ns/day dt: 1.5fs (ran 100000 steps in 17.84s)
```